pyproject.toml
[tool.poetry]
name = "blog"
version = "0.1.0"
description = ""
authors = ["AyushAryal <ayusharyal100@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^5.0.1"
django-ckeditor = "^6.7.0"
django-colorfield = "^0.11.0"
Pygments = "^2.17.2"
pillow = "^10.2.0"
requests = "^2.31.0"
markdown = "^3.5.2"
beautifulsoup4 = "^4.12.3"


[tool.poetry.group.dev.dependencies]
django-stubs = "^4.2.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
